Refactor UI control code.
