+++ src/control/ui/index.js
@@ -1,0 +1,2 @@
+export * from './FoundryUIManager.js';
+export * from './ViewManager.js';
