+++ src/view/log/HandlerLog.js
@@ -1,6 +1,8 @@
 import QuestAPI      from '../../control/public/QuestAPI.js';
 import QuestDB       from '../../control/QuestDB.js';
-import ViewManager   from '../../control/ViewManager.js';
+
+import { ViewManager }  from '../../control/index.js';
+
 import Socket        from '../../control/Socket.js';
 import FQLDialog     from '../FQLDialog.js';
